requirements.txt
cycler==0.11.0
dill==0.3.4
fonttools==4.31.2
groot-trees==0.0.16
gurobipy==9.1.2
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.0
matplotlib==3.5.1
numba==0.55.1
numpy==1.21.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.64.0
